+++ src/main/java/org/corpus_tools/annis/gui/ExportPanel.java
@@ -13,33 +13,39 @@
  */
 package org.corpus_tools.annis.gui;
 
+import com.google.common.base.Joiner;
+import com.google.common.base.Splitter;
 import com.google.common.base.Stopwatch;
 import com.google.common.eventbus.EventBus;
 import com.google.common.eventbus.Subscribe;
 import com.vaadin.data.Binder;
 import com.vaadin.data.HasValue;
+import com.vaadin.icons.VaadinIcons;
 import com.vaadin.server.FileDownloader;
 import com.vaadin.server.FileResource;
 import com.vaadin.server.FontAwesome;
 import com.vaadin.ui.Button;
 import com.vaadin.ui.Button.ClickEvent;
 import com.vaadin.ui.ComboBox.NewItemProvider;
+import com.vaadin.ui.Component;
+import com.vaadin.ui.CssLayout;
 import com.vaadin.ui.FormLayout;
 import com.vaadin.ui.GridLayout;
 import com.vaadin.ui.HorizontalLayout;
 import com.vaadin.ui.Notification;
 import com.vaadin.ui.ProgressBar;
+import com.vaadin.ui.TextField;
 import com.vaadin.ui.UI;
 import com.vaadin.ui.VerticalLayout;
+import com.vaadin.ui.themes.ValoTheme;
 import com.vaadin.v7.shared.ui.label.ContentMode;
 import com.vaadin.v7.ui.CheckBox;
 import com.vaadin.v7.ui.Label;
-import com.vaadin.v7.ui.TextField;
 import java.io.File;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Optional;
-import org.corpus_tools.annis.gui.components.HelpButton;
 import org.corpus_tools.annis.gui.controlpanel.SearchOptionsPanel;
-import org.corpus_tools.annis.gui.converter.CommaSeperatedStringConverterList;
 import org.corpus_tools.annis.gui.exporter.ExporterPlugin;
 import org.corpus_tools.annis.gui.objects.QueryUIState;
 import org.slf4j.LoggerFactory;
@@ -154,6 +160,30 @@ public class ExportPanel extends GridLayout {
 
   }
 
+  public static class HelpButtonListener implements Button.ClickListener {
+
+    private final Component component;
+
+    public HelpButtonListener(Component component) {
+      this.component = component;
+    }
+
+    @Override
+    public void buttonClick(ClickEvent event) {
+      String caption = "Help";
+      if (component.getCaption() != null && !component.getCaption().isEmpty()) {
+        caption = "Help for \"" + component.getCaption() + "\"";
+      }
+      caption = caption + "<br/><br/>(Click here to close)";
+      Notification notify = new Notification(caption, Notification.Type.HUMANIZED_MESSAGE);
+      notify.setHtmlContentAllowed(true);
+      notify.setDescription(component.getDescription());
+      notify.setDelayMsec(-1);
+      notify.show(UI.getCurrent().getPage());
+    }
+
+  }
+
   /**
    * 
    */
@@ -198,6 +228,12 @@ public class ExportPanel extends GridLayout {
   private final Label lblHelp;
 
   private final CheckBox cbAlignmc;
+
+
+  private static final Splitter paramSplitter = Splitter.on(',').trimResults();
+
+  private static final Joiner paramJoiner = Joiner.on(", ");
+
 
   public ExportPanel(AnnisUI ui) {
     super(2, 3);
@@ -260,17 +296,33 @@ public class ExportPanel extends GridLayout {
     formLayout.addComponent(cbLeftContext);
     formLayout.addComponent(cbRightContext);
 
-    txtAnnotationKeys = new TextField("Annotation Keys");
-    txtAnnotationKeys.setDescription("Some exporters will use this comma "
+    txtAnnotationKeys = new TextField();
+    Button btHelpAnnotationKeys = new Button("");
+    btHelpAnnotationKeys.setIcon(VaadinIcons.QUESTION);
+    btHelpAnnotationKeys.addStyleName(ValoTheme.BUTTON_BORDERLESS);
+    btHelpAnnotationKeys.addStyleName("helpbutton");
+
+    CssLayout keysLayout = new CssLayout(txtAnnotationKeys, btHelpAnnotationKeys);
+    keysLayout.setCaption("Annotation Keys");
+    keysLayout.setDescription("Some exporters will use this comma "
         + "seperated list of annotation keys to limit the exported data to these "
         + "annotations.");
-    formLayout.addComponent(new HelpButton<String>(txtAnnotationKeys));
+    btHelpAnnotationKeys.addClickListener(new HelpButtonListener(keysLayout));
+    formLayout.addComponent(keysLayout);
 
-    txtParameters = new TextField("Parameters");
-    txtParameters.setDescription("You can input special parameters "
+    txtParameters = new TextField();
+    Button btHelpParameters = new Button("");
+    btHelpParameters.setIcon(VaadinIcons.QUESTION);
+    btHelpParameters.addStyleName(ValoTheme.BUTTON_BORDERLESS);
+    btHelpParameters.addStyleName("helpbutton");
+    CssLayout parametersLayout = new CssLayout(txtParameters, btHelpParameters);
+    parametersLayout.setCaption("Parameters");
+    parametersLayout.setDescription("You can input special parameters "
         + "for certain exporters. See the description of each exporter "
         + "(‘?’ button above) for specific parameter settings.");
-    formLayout.addComponent(new HelpButton<String>(txtParameters));
+
+    btHelpParameters.addClickListener(new HelpButtonListener(parametersLayout));
+    formLayout.addComponent(parametersLayout);
 
     // check box for match-with-context exporter
     cbAlignmc = new CheckBox("align matches" + "<br/>" + "by node number");
@@ -327,10 +379,16 @@ public class ExportPanel extends GridLayout {
       cbLeftContext.addSelectionListener(event -> binder.setBean(state));
       cbRightContext.addSelectionListener(event -> binder.setBean(state));
 
-      txtAnnotationKeys.setConverter(new CommaSeperatedStringConverterList());
-      txtAnnotationKeys.setPropertyDataSource(state.getExportAnnotationKeys());
 
-      txtParameters.setPropertyDataSource(state.getExportParameters());
+      binder.forField(txtAnnotationKeys).withConverter(presentation -> {
+        List<String> result = new ArrayList<>();
+        for (String s : paramSplitter.split(presentation)) {
+          result.add(s);
+        }
+        return result;
+      }, paramJoiner::join).bind("exportAnnotationKeys");
+
+      binder.forField(txtParameters).bind("exportParameters");
 
       cbAlignmc.setPropertyDataSource(state.getAlignmc());
 
@@ -385,8 +443,10 @@ public class ExportPanel extends GridLayout {
     // when not longer needed
     tmpOutputFile = currentTmpFile;
     //
-    if (exportError instanceof IllegalStateException || exportError instanceof ClassCastException) {
-      Notification.show(exportError.getMessage(), Notification.Type.ERROR_MESSAGE);
+    if (exportError instanceof IllegalStateException
+        || exportError instanceof IllegalArgumentException
+        || exportError instanceof ClassCastException) {
+      Notification.show("Export failed", exportError.getMessage(), Notification.Type.ERROR_MESSAGE);
     } else if (tmpOutputFile == null) {
       Notification.show("Could not create the Exporter",
           "The server logs might contain more information about this "