+++ CHANGELOG.md
@@ -5,6 +5,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Show informative error message when segmentation in export does not exist
+  instead of a generic stack trace
+
 ## [4.9.9] - 2022-10-26
 
 ### Fixed